refactor(chat): neutralize run-state helper service filenames
