+++ turbo/apps/api/src/signals/services/zero-goal.service.ts
@@ -18,7 +18,7 @@ import { publishChatThreadMessageCreatedSafely } from "../external/realtime";
 import {
   appendGoalCloseMarker,
   appendGoalOpenMarker,
-} from "./zero-chat-goal-marker.service";
+} from "./chat-goal-marker.service";
 import { normalizeGoalObjectiveBrief } from "./goal-objective-brief-normalization.service";
 import { generateGoalObjectiveBrief } from "./zero-goal-objective-brief.service";
 import { lockGoalThread } from "./goal-lock.service";
